pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kryoset"
version = "1.0.0"
description = "A secure self-hosted NAS server over SFTP."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "paramiko>=3.4.0",
    "bcrypt>=4.1.0",
    "click>=8.1.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.2",
    "Pillow>=10.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "cryptography>=42.0.0",
    "jinja2>=3.1.0",
]

[project.scripts]
kryoset = "kryoset.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["kryoset/tests"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
where = ["."]
include = ["kryoset*"]
